Merged
Closes #19
