+++ packages/core/src/service/openapi.ts
@@ -1533,7 +1533,7 @@ export class OpenApiService {
     try {
       const response = await fetch(`https://rest-${network}.onflow.org/v1/blocks?height=sealed`);
       const result = await response.json();
-      return result[0].header !== null && result[0].header !== undefined;
+      return Array.isArray(result) && result[0].header !== null && result[0].header !== undefined;
     } catch (err) {
       return false;
     }
