--- packages/core/src/utils/modules/settings.ts
@@ -1,93 +1,0 @@
-import { getLocalData, setLocalData, removeLocalData } from '@onflow/frw-data-model';
-
-import { consoleError } from '@onflow/frw-shared/utils';
-
-import { jsonToString } from './utils';
-
-type CredentialSetting = {
-  instant: string;
-  user: {
-    name: string;
-    displayName: string;
-  };
-  id: string;
-  credentialPublicKey: Uint8Array;
-  response: AuthenticatorAttestationResponse;
-};
-type Settings = {
-  credentials: Record<string, CredentialSetting>;
-  rp?: Record<string, unknown>;
-  user?: Record<string, unknown>;
-};
-
-export function addCredential(
-  settings: Settings,
-  user: PublicKeyCredentialUserEntity,
-  id: string,
-  credentialPublicKey: Uint8Array,
-  response: AuthenticatorAttestationResponse
-) {
-  if (id && id.length > 0) {
-    settings.credentials[id] = {
-      instant: new Date().toISOString(),
-      user: {
-        name: user.name,
-        displayName: user.displayName,
-      },
-      id: id,
-      credentialPublicKey: credentialPublicKey,
-      response: response,
-    };
-    saveSettings(settings);
-  }
-}
-
-export function getCredential(settings: Settings, id: string) {
-  const cred = settings.credentials[id];
-  return cred;
-}
-
-export async function getUsername(id: string): Promise<string | null> {
-  const settings = await readSettings();
-  if (id in settings.credentials) {
-    const cred = settings.credentials[id];
-    return cred.user.name;
-  }
-  return null;
-}
-
-export async function readSettings(): Promise<Settings> {
-  let settings: Settings = {
-    credentials: {},
-  };
-  const s = await getLocalData<string>('passkey-settings');
-  if (!s) {
-    return settings;
-  }
-
-  try {
-    settings = JSON.parse(s);
-
-    if (!settings.credentials) {
-      settings.credentials = {};
-    }
-    if ('rp' in settings) {
-      delete settings.rp;
-    }
-    if ('user' in settings) {
-      delete settings.user;
-    }
-  } catch {
-    consoleError('Error parsing settings');
-  }
-
-  return settings;
-}
-
-export function saveSettings(settings: Settings) {
-  if (settings) {
-    setLocalData('passkey-settings', jsonToString(settings));
-  } else {
-    removeLocalData('passkey-settings');
-  }
-}
