+++ packages/core/tsconfig.json
@@ -1,8 +1,5 @@
 {
   "extends": "../../tsconfig.base.json",
-  "compilerOptions": {
-    "lib": ["ES2022", "DOM"]
-  },
   "references": [
     {
       "path": "../shared"
